build-cross: clean up warnings exposed under the QEMU cross matrix.
A handful of diagnostics raised across the 12-arch cross build that
the host gcc/clang on macOS / x86 didn't fire. Each fix is narrowly
scoped:

  ssfassert.h    add SSF_ASSUME(x) optimizer hint
                 (__builtin_unreachable / __assume). Pairs with
                 SSF_REQUIRE when the same precondition also has
                 to drive downstream value-range analysis past
                 inlining.

  ssfbn.c        SSF_ASSUME the SSFBNCopy length bound so the
                 trailing memcpy clears -Warray-bounds when the
                 inliner loses the SSF_REQUIRE'd invariant.

  ssfbn_ut.c     cast BN_num_bytes(src) result to size_t for the
                 bound-check comparison (sign-compare).

  ssfec.c        _SSFECScalarMulVTwNAF: loop counters int32_t ->
                 size_t; the descending digit loop is rewritten
                 as i-- > 0 to drop the signed-bound compare.

  ssfasn1.c      zero-init the IntU64 scratch buffer.

  ssfecdsa_ut.c  pin OPENSSL_API_COMPAT to 10101 in this TU only.
                 The test deliberately uses the legacy EC_KEY /
                 ECDSA_do_* / ECDH_compute_key APIs as a verifier
                 oracle; OpenSSL 3.0+ flags those for deprecation.
                 Production code keeps -Wdeprecated-declarations
                 active.

  ssfed25519_ut.c, ssfrsa_ut.c, ssfx25519_ut.c
                 wrap the stack-scanning helpers in a pragma
                 push/pop disabling -Wuninitialized /
                 -Wmaybe-uninitialized. Reading the uninit prefix
                 is the test (each helper verifies post-call
                 stack hygiene).

  ssfhkdf_ut.c   replace the OSSL digest-name copy loop with
                 strlen + SSF_ASSERT + memcpy + null-terminate.

Verified: ./tools/cross-test/run.sh --docker all -> 12/12 PASS
(aarch64, armhf, i686, m68k, mips, mips64, mips64el, mipsel,
mipsn32, mipsn32el, powerpc, riscv64).

+++ _crypto/ssfbn_ut.c
@@ -80,7 +80,7 @@ static void _FromOSSL(SSFBN_t *dst, const BIGNUM *src, uint16_t numLimbs)
     size_t outSize = (size_t)numLimbs * sizeof(SSFBNLimb_t);
 
     SSF_ASSERT(outSize <= sizeof(buf));
-    SSF_ASSERT(BN_num_bytes(src) <= (int)outSize);
+    SSF_ASSERT((size_t)BN_num_bytes(src) <= outSize);
     SSF_ASSERT(BN_bn2binpad(src, buf, (int)outSize) == (int)outSize);
     SSF_ASSERT(SSFBNFromBytes(dst, buf, outSize, numLimbs) == true);
 }
